Implement Orleans stream background service and INotificationClient interface for improved message handling
