+++ samples/MessageEachOther/MessageEachOther.ApiService/INotificationClient.cs
@@ -1,0 +1,6 @@
+namespace MessageEachOther.ApiService;
+
+interface INotificationClient
+{
+    Task ReceiveMessage(string request);
+}
